target-bigquery: Add placeholder for the Location
Having tried both US and EU as values with corresponding BigQuery datasets,
I found that neither worked and should point to the `region` as linked to. Without
setting the value to a good region the target requests an API endpoint that returns
a 404 status code not making clear the region/location might be misconfigured.